Add optional config field `TestFolder`
If users want to have the tests being run in the pipeline somewhere
different than in the `test` folder, they can specify a `TestFolder`
in the config.yaml.
This might be useful, if developers want to have `unittests` in the
`test` folder and the tests for the pipeline in the subfolder of
`test`.

An example:
```
|test
|- unittests
|- integration_tests
 |- test_test1.py
```

If the user wants to run `test_test1.py` in the pipeline,
the entry in the config.yaml should be
`TestFolder: test/integration_tests`